+++ CHANGELOG.md
@@ -4,3 +4,4 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 - Start of new changelog.
+- Fixed GM SB XRAY button using stored order ID when none found in email.